+++ src/include/evaluator_inc.hpp
@@ -52,6 +52,7 @@ class IncrementalEvaluator {
   bool checkLoopBody();
   bool checkPostLoop();
   bool isCommutative(int64_t cell) const;
+  bool isCommutative(const std::set<int64_t>& cells) const;
   void computeStatefulCells();
   void computeLoopCounterDependentCells();