+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Bugfixes
+
+* Fix incorrect result of incremental evaluator
+
 ### Enhancements
 
 * Support more recursive formulas, e.g. [A000142](https://loda-lang.org/edit/?oeis=142)